Guard removal of proxy file in unit test

+++ plugin_tests/proxy_test.py
@@ -21,7 +21,8 @@ class ProxyTestCase(base.TestCase):
         super(ProxyTestCase, self).setUp()
 
         self._proxy_file_path = '/tmp/proxy'
-        os.remove('%s.db' % self._proxy_file_path)
+        if os.path.exists(self._proxy_file_path):
+            os.remove('%s.db' % self._proxy_file_path)
 
         users = ({
             'email': 'cumulus@email.com',
